# src/api/openapi/Heimatplatz.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Heimatplatz.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/forgot-password": {
      "post": {
        "tags": [
          "Heimatplatz.Api"
        ],
        "operationId": "ForgotPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Heimatplatz.Api"
        ],
        "operationId": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Heimatplatz.Api"
        ],
        "operationId": "RefreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Heimatplatz.Api"
        ],
        "operationId": "Register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/resend-verification": {
      "post": {
        "tags": [
          "Heimatplatz.Api"
        ],
        "operationId": "ResendVerificationEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResendVerificationEmailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "tags": [
          "Heimatplatz.Api"
        ],
        "operationId": "ResetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "tags": [
          "Heimatplatz.Api"
        ],
        "operationId": "VerifyEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ForgotPasswordRequest": {
        "required": [
          "Email"
        ],
        "type": "object",
        "properties": {
          "Email": {
            "type": "string"
          }
        }
      },
      "ForgotPasswordResponse": {
        "required": [
          "Message"
        ],
        "type": "object",
        "properties": {
          "Message": {
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "required": [
          "Email",
          "Password"
        ],
        "type": "object",
        "properties": {
          "Email": {
            "type": "string"
          },
          "Password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "required": [
          "AccessToken",
          "RefreshToken",
          "UserId",
          "Email",
          "FullName",
          "ExpiresAt"
        ],
        "type": "object",
        "properties": {
          "AccessToken": {
            "type": "string"
          },
          "RefreshToken": {
            "type": "string"
          },
          "UserId": {
            "type": "string",
            "format": "uuid"
          },
          "Email": {
            "type": "string"
          },
          "FullName": {
            "type": "string"
          },
          "ExpiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "RefreshTokenRequest": {
        "required": [
          "RefreshToken"
        ],
        "type": "object",
        "properties": {
          "RefreshToken": {
            "type": "string"
          }
        }
      },
      "RefreshTokenResponse": {
        "required": [
          "AccessToken",
          "RefreshToken",
          "ExpiresAt"
        ],
        "type": "object",
        "properties": {
          "AccessToken": {
            "type": "string"
          },
          "RefreshToken": {
            "type": "string"
          },
          "ExpiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "RegisterRequest": {
        "required": [
          "FirstName",
          "LastName",
          "Email",
          "Password"
        ],
        "type": "object",
        "properties": {
          "FirstName": {
            "type": "string"
          },
          "LastName": {
            "type": "string"
          },
          "Email": {
            "type": "string"
          },
          "Password": {
            "type": "string"
          },
          "SellerType": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SellerType"
              }
            ]
          },
          "CompanyName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RegisterResponse": {
        "required": [
          "AccessToken",
          "RefreshToken",
          "UserId",
          "Email",
          "FullName",
          "ExpiresAt"
        ],
        "type": "object",
        "properties": {
          "AccessToken": {
            "type": "string"
          },
          "RefreshToken": {
            "type": "string"
          },
          "UserId": {
            "type": "string",
            "format": "uuid"
          },
          "Email": {
            "type": "string"
          },
          "FullName": {
            "type": "string"
          },
          "ExpiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ResendVerificationEmailRequest": {
        "type": "object"
      },
      "ResendVerificationEmailResponse": {
        "type": "object",
        "properties": {
          "AlreadyVerified": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "ResetPasswordRequest": {
        "required": [
          "Token",
          "NewPassword"
        ],
        "type": "object",
        "properties": {
          "Token": {
            "type": "string"
          },
          "NewPassword": {
            "type": "string"
          }
        }
      },
      "ResetPasswordResponse": {
        "required": [
          "Message"
        ],
        "type": "object",
        "properties": {
          "Message": {
            "type": "string"
          }
        }
      },
...
      "SellerType": {
        "enum": [
          "Private",
          "Broker",
          "PropertyManager",
          null
        ]
      },
...
      "VerifyEmailRequest": {
        "required": [
          "Token"
        ],
        "type": "object",
        "properties": {
          "Token": {
            "type": "string"
          }
        }
      },
      "VerifyEmailResponse": {
        "required": [
          "Email"
        ],
        "type": "object",
        "properties": {
          "Email": {
            "type": "string"
          },
          "AlreadyVerified": {
            "type": "boolean",
            "default": false
          }
        }
      }
    }
  },
...
}